Agregada mas funciones para la subida a mega y borrado de los archivos

+++ index.js
@@ -58,35 +58,78 @@ bot.on("document", async(ctx) => {
     
     let random = (Math.random()*1000*document.file_size / (1024 * 1024)).toString().replace('.', '');
 
-    ctx.reply(`Downloading file: ${nombre} ${(document.file_size / (1024 * 1024)).toFixed(2)}MB`);
+    ctx.reply(`Downloading file: \`${nombre}\` ${(document.file_size / (1024 * 1024)).toFixed(2)}MB`, {parse_mode: "Markdown"});
     await download(file, process.env.OUT_DIR+random);
     fs.renameSync(process.env.OUT_DIR+random+'/'+file.href.split("/").pop().replace(file.search, ''), process.env.OUT_DIR+nombre);
     fs.rmdirSync(process.env.OUT_DIR+random);
-    ctx.reply(`Downloaded file: ${document.file_name} ${(document.file_size / (1024 * 1024)).toFixed(2)}MB`);
+    ctx.reply(`Downloaded file: \`${document.file_name}\` ${(document.file_size / (1024 * 1024)).toFixed(2)}MB`, {parse_mode: "Markdown"});
 })
 
-bot.command("uploadMegaAllDelete", async(ctx) => {
-    // MEGA LOGIN
-    const storage = await new Storage({
+const megaLogin = async() => {
+    return await new Storage({
         email: process.env.MEGA_EMAIL,
         password: process.env.MEGA_PASSWORD
     }).ready
+}
 
+const uploadToMega = async(storage, file, ctx) => {
+    ctx.reply(`Uploading to MEGA: \`${path.basename(file)}\``, {parse_mode: "Markdown"});
+    await storage.upload({
+        name: path.basename(file),
+        size: fs.statSync(file).size
+    }, fs.createReadStream(file)).complete
+    ctx.reply('The file was uploaded!');
+    return;
+}
+
+const uploadMega = async(ctx, deleteFile) => {
+    // MEGA LOGIN
+    const storage = await megaLogin();
+
+    // UPLOAD FILE
+    let fileName = "";
+    const chunks = ctx.message.text.split(" ");
+    chunks.shift();
+    for (let i = 0; i < chunks.length; i++) {
+        fileName += `${chunks[i]}${(i!=chunks.length-1)?" ":""}`;
+    }
+
+    let file = process.env.OUT_DIR + fileName;
+    if (fs.existsSync(file) && fs.statSync(file).isFile()) {
+        await uploadToMega(storage, file, ctx)
+        if (deleteFile) fs.rmSync(file);
+    } else {
+        ctx.reply("undefined file");
+    }
+}
+
+const uploadMegaAll = async(ctx, deleteFile) => {
+    // MEGA LOGIN
+    const storage = await megaLogin();
+    
     // UPLOADING FILES
     let dir = fs.readdirSync(process.env.OUT_DIR);
     for (const element of dir) {
         let file = process.env.OUT_DIR + element;
-        ctx.reply(`Uploading to MEGA: ${path.basename(file)}`);
-        await storage.upload({
-            name: path.basename(file),
-            size: fs.statSync(file).size
-        }, fs.createReadStream(file)).complete
-        ctx.reply('The file was uploaded!');
-        fs.rmSync(file);
+        await uploadToMega(storage, file, ctx)
+        if (deleteFile) fs.rmSync(file);
     }
+}
+
+
+bot.command("uploadMega", async(ctx) => await uploadMega(ctx, false));
+bot.command("uploadMegaDelete", async(ctx) => await uploadMega(ctx, true));
+bot.command("uploadMegaAll", async(ctx) => await uploadMegaAll(ctx, false));
+bot.command("uploadMegaAllDelete", async(ctx) => await uploadMegaAll(ctx, true));
+
+bot.hears("ls", async(ctx) => {
+    let dir = fs.readdirSync(process.env.OUT_DIR);
+    let message = `The output folder has the following files:`;
+    for (const file of dir) {
+        message += `\n\`${file}\` (${(fs.statSync(process.env.OUT_DIR+file).size / (1024 * 1024)).toFixed(2)}MB)`;
+    }
+    ctx.reply(message, {parse_mode: "Markdown"})
 })
-
-
 
 bot.launch()
 console.log("Iniciando Violet")
